% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_class.R
\name{assign_class}
\alias{assign_class}
\title{Assigning S3 classes for rendering}
\usage{
assign_class(x)
}
\arguments{
\item{x}{A character vector or a list.}
}
\description{
This function assess the structure of input vectors or lists in order to use
the proper rendering method by \code{\link[=write_yaml]{write_yaml()}}.
}
\keyword{internal}
