% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print2text}
\alias{print2text}
\title{Print content of rmd_doc into a text string}
\usage{
print2text(x)
}
\arguments{
\item{x}{Input \linkS4class{rmd_doc} object.}
}
\description{
Content of \linkS4class{rmd_doc} objects will be converted into a string that will
be passed either to \code{\link[=print]{print()}} or to \code{\link[=write_rmd]{write_rmd()}}.
}
\keyword{internal}
