% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_yaml}
\alias{write_yaml}
\title{Writing yaml headers}
\usage{
write_yaml(..., append, body = "", filename)
}
\arguments{
\item{...}{Named arguments to be inserted in the yaml head. At the moment
only vectors are implemented (length 1 or higher), but the concept may
also consider lists for complex tags.}

\item{append}{A piece of code to be appended in the head. For instance, it
can contain commented code.}

\item{body}{The content of the document that will be inserted after the head.}

\item{filename}{A character value with the name of the file to be written
(usually a *.Rmd file). If missing, no file will be written by the
function.}
}
\description{
This function generates yaml headers and Rmarkdown files by including the
settings as arguments of the function.
Comments and pieces of header can be also added through the argument
\code{append}.
}
