% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as}
\alias{as}
\alias{coerce,list,rmd_doc-method}
\title{Coercing lists and rmd_doc objects}
\description{
Objects of class \linkS4class{rmd_doc} can be created from lists or converted back
to lists. This is convenient for defining functions that manipulate the
content of such objects.
}
