% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rmd.R
\name{write_rmd}
\alias{write_rmd}
\title{Writing R-Markdown Documents}
\usage{
write_rmd(..., append, body = "", filename)
}
\arguments{
\item{...}{Named arguments to be inserted in the YAML header. These arguments
may be inserted either as vectors or lists.}

\item{append}{A piece of code to be appended in the header. For instance, it
can contain commented code.}

\item{body}{The content of the document that will be inserted after the
header.}

\item{filename}{A character value with the name of the file to be written.
If not included, the extension *.Rmd will be appended to this name.
If missing, no file will be written by this function.}
}
\value{
A character vector of class \code{rmd_doc} and, if argument set for parameter
\code{filename}, an Rmd file.
}
\description{
This function generates YAML headers and R-Markdown documents by including
the settings as arguments of the function.
Comments and pieces of header can be also added through the argument
\code{append}.
}
\examples{
\dontrun{
my_document <- write_rmd(
  title = "Sample Document", author = "Miguel Alavarez",
  output = "html_document",
  body = txt_body(
    "# Intro",
    "",
    "This is just an example."
  )
)
my_document
}

}
