% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_yaml}
\alias{write_yaml}
\title{Write yaml entries}
\usage{
write_yaml(x, ...)
}
\arguments{
\item{x}{Either a vector or a list.}

\item{...}{Further arguments passed among methods.}

\item{depth}{Hierarchical depth in lists.}
}
\description{
Internal function for a proper rendering of yaml entry classes as arguments
in \code{\link[=write_rmd]{write_rmd()}}.
}
\keyword{internal}
